make bootstrap explicit js-file
